git-svn-id: https://svn.lcsr.jhu.edu/cisst/trunk@3811 0959eb1a-6956-4f60-947b-6092a4e22c97

+++ cisst/cisstMultiTask/code/mtsComponentAddLatency.cpp
@@ -407,7 +407,7 @@ double mtsComponentAddLatency::SetLatency(double latencyInSeconds)
 {
     double previousLatency = this->Latency;
     this->Latency = latencyInSeconds;
-    const size_t latencyInPeriods = latencyInSeconds / this->GetPeriodicity();
+    const size_t latencyInPeriods = static_cast<size_t>(latencyInSeconds / this->GetPeriodicity());
     LatencyStateTable.SetDelay(latencyInPeriods);
     return previousLatency;
 }
